FInalisation de auth, suppression du compte en cours de dev

+++ serveur/push.php
@@ -32,8 +32,8 @@ function pushCousesIndependent($user, PDO $bdd){
 
 		return checkGroupe($user, $bdd, function($user, $groupe, $bdd){
 	
-			if(isset($_POST['submitCourse'])) {
-				$titre = htmlspecialchars( $_POST['titre']);
+			if(isset($_POST['submitCourse']) && isset($_POST['titre']) && isset($_POST['maxPrice'])) {
+				$titre = htmlspecialchars($_POST['titre']);
 				$maxPrice = (float)  $_POST['maxPrice'];
 		
 				$insert = $bdd->prepare('INSERT INTO courses (`nom`, `maxprice`, `groupe`) VALUES (?,?,?)');
@@ -91,8 +91,8 @@ function pushCousesIndependent($user, PDO $bdd){
 
 function push($user, $usedCourse, PDO $bdd){
 
-	if (isset($_POST['submitArticle'])) {
-		$titre = htmlspecialchars((string) $_POST['titre']);
+	if (isset($_POST['submitArticle']) && isset($_POST['titre']) && isset($_POST['prix'])) {
+		$titre = htmlspecialchars($_POST['titre']);
 		$prix = (float) $_POST['prix'];
 
 		$insertArt = $bdd->prepare('INSERT INTO articles (titre, prix, course, preview, idAuteur) VALUES (?,?,?,0,?)');
@@ -105,6 +105,7 @@ function push($user, $usedCourse, PDO $bdd){
 		$reqIdInserted->execute(array($usedCourse['id']));
 		$idInserted = $reqIdInserted->fetch();
 		echo json_encode([
+			'status' => 200,
 			'id' => $idInserted['id'],
 			'titre' => $titre,
 			'prix' => $prix
@@ -112,8 +113,8 @@ function push($user, $usedCourse, PDO $bdd){
 
 		return true;
 	}
-	elseif (isset($_POST['submitPreview'])) {
-		$titre = htmlspecialchars((string) $_POST['titre']);
+	elseif (isset($_POST['submitPreview']) && isset($_POST['titre'])) {
+		$titre = htmlspecialchars($_POST['titre']);
 
 		$insertArt = $bdd->prepare('INSERT INTO articles (titre, prix, course, preview, idAuteur) VALUES (?,0,?,1,?)');
 		$insertArt->execute(array($titre, $usedCourse['id'], $user['id']));
@@ -123,6 +124,7 @@ function push($user, $usedCourse, PDO $bdd){
 		$reqIdInserted->execute(array($usedCourse['id']));
 		$idInserted = $reqIdInserted->fetch();
 		echo json_encode([
+			'submit' => 200,
 			'id'=> $idInserted['id'],
 			'titre'=> $titre,
 			'color' => $user['hexColor']
@@ -130,7 +132,7 @@ function push($user, $usedCourse, PDO $bdd){
 		
 		return true;
 	}
-	elseif (isset($_POST['deleteCourse'])) {
+	elseif (isset($_POST['deleteCourse']) && $_POST['id']) {
 
 		$reqDelete = $bdd->prepare('DELETE FROM `courses` WHERE `id` = :index');
 		$reqDelete->bindParam(':index', $usedCourse['id'], PDO::PARAM_INT);
@@ -142,47 +144,45 @@ function push($user, $usedCourse, PDO $bdd){
 		echo json_encode(array('status' => 200));
 		return true;
 	}
-	elseif (isset($_POST['deleteArticle'])) {
-		$_POST['id'] = (int) $_POST['id'];
+	elseif (isset($_POST['deleteArticle']) && isset($_POST['id'])) {
 
-		$reqIdDeleted = $bdd->prepare('SELECT prix FROM articles WHERE id = ?');
-		$reqIdDeleted->execute(array((int) $_POST['id']));
-		$idDeleted = $reqIdDeleted->fetch();
+		$reqDeleted = $bdd->prepare('SELECT `id`, `prix` FROM `articles` WHERE `id` = ?');
+		$reqDeleted->execute(array($_POST['id']));
+		$deleted = $reqDeleted->fetch();
 
 		$reqDelete = $bdd->prepare('DELETE FROM articles WHERE id=:index');
-		$reqDelete->bindParam(':index', $_POST['id'], PDO::PARAM_INT);
+		$reqDelete->bindParam(':index', $deleted['id'], PDO::PARAM_INT);
 		$reqDelete->execute();
 
 		$updCourse = $bdd->prepare('UPDATE courses SET total=? WHERE id=?');
-		$updCourse->execute(array($usedCourse['total'] - $idDeleted['prix'], $usedCourse['id']));
+		$updCourse->execute(array($usedCourse['total'] - $deleted['prix'], $usedCourse['id']));
 
-		echo json_encode(['done', $idDeleted['prix']]);
+		echo json_encode(array('status' => 200, 'prix' => $deleted['prix']));
 		
 		return true;
 	}
-	elseif (isset($_POST['deletePreview'])) {
-		$_POST['id'] = (int) $_POST['id'];
-
+	elseif (isset($_POST['deletePreview']) && isset($_POST['id'])) {
 		$reqDelete = $bdd->prepare('DELETE FROM articles WHERE id=:index');
 		$reqDelete->bindParam(':index', $_POST['id'], PDO::PARAM_INT);
 		$reqDelete->execute();
 
-		echo json_encode(['done']);
+		echo json_encode(array('status' => 200));
 		
 		return true;
 	}
-	elseif(isset($_POST['buyPreview'])) {
+	elseif(isset($_POST['buyPreview']) && isset($_POST['id']) && isset($_POST['prix'])) {
 		$updCourse = $bdd->prepare('UPDATE articles SET preview=0, prix=? WHERE id=?');
-		$updCourse->execute(array((float) $_POST['prix'],(int) $_POST['id']));
+		$updCourse->execute(array($_POST['prix'], $_POST['id']));
 
 		$updCourse = $bdd->prepare('UPDATE courses SET total=? WHERE id=?');
-		$updCourse->execute(array($usedCourse['total'] + (float) $_POST['prix'], $usedCourse['id']));
+		$updCourse->execute(array($usedCourse['total'] + $_POST['prix'], $usedCourse['id']));
 
 		$reqIdInserted = $bdd->prepare('SELECT * FROM articles WHERE id=?');
-		$reqIdInserted->execute(array((int) $_POST['id']));
+		$reqIdInserted->execute(array($_POST['id']));
 		$idInserted = $reqIdInserted->fetch();
 
 		echo json_encode([
+			'status' => 200,
 			'id' => $_POST['id'],
 			'titre' => $idInserted['titre'],
 			'prix' => $idInserted['prix']
